Switch to `AbstractBundle`
